==> src/Markdown/02_Genome_catalogue_overview.Rmd <==
---
title: "Genome catalogue overview"
output: html_document
date: "2024-01-19"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r libraries, comment="", echo=FALSE, message=FALSE, warning=FALSE}
library(tidyverse)
library(ape)
library(distillR)
library(ggtree)
library(ggnewscale)
library(ggtreeExtra)
library(gridExtra)
library(phytools)
library(Rtsne)
```


```{r loaddata, comment="", echo=FALSE, message=FALSE, warning=FALSE}
load("../../resources/data/data.Rdata")
```

# Genome phylogeny
```{r colors, comment="", echo=FALSE, message=FALSE, warning=FALSE, results='hide'}
# Get phylum colors from the EHI standard
phylum_colors <- read_tsv("https://raw.githubusercontent.com/earthhologenome/EHI_taxonomy_colour/main/ehi_phylum_colors.tsv") %>%
  mutate_at(vars(phylum), ~ str_replace(., "[dpcofgs]__", "")) %>%
  right_join(genome_metadata, by = join_by(phylum == phylum)) %>%
  arrange(match(genome, bacteria_tree$tip.label)) %>%
  select(phylum, colors) %>%
  unique() %>%
  arrange(phylum) %>%
  select(colors) %>%
  pull()

# Generate the phylum color heatmap
phylum_heatmap <- read_tsv("https://raw.githubusercontent.com/earthhologenome/EHI_taxonomy_colour/main/ehi_phylum_colors.tsv") %>%
  mutate_at(vars(phylum), ~ str_replace(., "[dpcofgs]__", "")) %>%
  right_join(genome_metadata, by = join_by(phylum == phylum)) %>%
  arrange(match(genome, bacteria_tree$tip.label)) %>%
  select(genome, phylum) %>%
  mutate(phylum = factor(phylum, levels = unique(phylum))) %>%
  column_to_rownames(var = "genome")

# Generate  basal tree
circular_tree <- bacteria_tree %>%
  force.ultrametric(method = "extend", message = FALSE) %>% # extend to ultrametric for the sake of visualisation
  ggtree(layout = "circular", size = 0.3, angle = 45)
```

```{r circular_tree, comment="", echo=FALSE, message=FALSE, warning=FALSE}
# Add phylum ring
circular_tree <- gheatmap(circular_tree, phylum_heatmap, offset = 0.85, width = 0.1, colnames = FALSE) +
  scale_fill_manual(values = phylum_colors) +
  geom_tiplab2(size = 1, hjust = -0.1) +
  theme(legend.position = "none", plot.margin = margin(0, 0, 0, 0), panel.margin = margin(0, 0, 0, 0))

# Flush color scale to enable a new color scheme in the next ring
circular_tree <- circular_tree + new_scale_fill()

# Add completeness ring
circular_tree <- circular_tree +
  new_scale_fill() +
  scale_fill_gradient(low = "#d1f4ba", high = "#f4baba") +
  geom_fruit(
    data = genome_metadata,
    geom = geom_bar,
    mapping = aes(x = Completeness, y = genome, fill = Contamination),
    offset = 0.55,
    orientation = "y",
    stat = "identity"
  )

# Add genome-size ring
circular_tree <- circular_tree +
  new_scale_fill() +
  scale_fill_manual(values = "#cccccc") +
  geom_fruit(
    data = genome_metadata,
    geom = geom_bar,
    mapping = aes(x = Genome_Size, y = genome),
    offset = 0.05,
    orientation = "y",
    stat = "identity"
  )

# Plot circular tree
circular_tree
```

# Genome quality

```{r biplot, comment="", echo=FALSE, message=FALSE, warning=FALSE}
# Generate quality biplot
genome_biplot <- genome_metadata %>%
  select(c(genome, domain, phylum, Completeness, Contamination, Genome_Size)) %>%
  arrange(match(genome, rev(bacteria_tree$tip.label))) %>% # sort MAGs according to phylogenetic tree
  ggplot(aes(x = Completeness, y = Contamination, size = Genome_Size, color = phylum)) +
  geom_point(alpha = 0.7) +
  ylim(c(10, 0)) +
  scale_color_manual(values = phylum_colors) +
  labs(y = "Contamination", x = "Completeness") +
  theme_classic() +
  theme(legend.position = "none")

# Generate contamination boxplot
genome_contamination <- genome_metadata %>%
  ggplot(aes(y = Contamination)) +
  ylim(c(10, 0)) +
  geom_boxplot(colour = "#999999", fill = "#cccccc") +
  theme_void() +
  theme(
    legend.position = "none",
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    plot.margin = unit(c(0, 0, 0.40, 0), "inches")
  ) # add bottom-margin (top, right, bottom, left)

# Generate completeness boxplot
genome_completeness <- genome_metadata %>%
  ggplot(aes(x = Completeness)) +
  xlim(c(50, 100)) +
  geom_boxplot(colour = "#999999", fill = "#cccccc") +
  theme_void() +
  theme(
    legend.position = "none",
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    plot.margin = unit(c(0, 0, 0, 0.50), "inches")
  ) # add left-margin (top, right, bottom, left)

# Render composite figure
grid.arrange(
  grobs = list(genome_completeness, genome_biplot, genome_contamination),
  layout_matrix = rbind(
    c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 4),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3),
    c(2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3)
  )
)
```

# Functional overview
```{r funcitonal, comment="", echo=FALSE, message=FALSE, warning=FALSE, results='hide'}
# Aggregate basal GIFT into elements
function_table <- genome_gifts %>%
  to.elements(., GIFT_db)

# Generate  basal tree
function_tree <-
  bacteria_tree %>%
  force.ultrametric(method = "extend", message = FALSE) %>%
  ggtree(size = 0.3)
```

```{r functional_tree, comment="", echo=FALSE, message=FALSE, warning=FALSE}
# Add phylum colors next to the tree tips
function_tree <-
  function_tree %>%
  gheatmap(phylum_heatmap, offset = 0, width = 0.1, colnames = FALSE) +
  scale_fill_manual(values = phylum_colors) +
  labs(fill = "Phylum")

# Reset fill scale to use a different colour profile in the heatmap
function_tree <- function_tree + new_scale_fill()

# Add functions heatmap
function_tree <- gheatmap(function_tree, function_table, offset = 0.5, width = 3.5, colnames = FALSE) +
  vexpand(.08) +
  coord_cartesian(clip = "off") +
  scale_fill_gradient(low = "#f4f4f4", high = "steelblue", na.value = "white") +
  labs(fill = "Function fullness")

# Reset fill scale to use a different colour profile in the heatmap
function_tree <- function_tree + new_scale_fill()

# Add completeness barplots
function_tree <- function_tree +
  geom_fruit(
    data = genome_metadata,
    geom = geom_bar,
    grid.params = list(axis = "x", text.size = 2, nbreak = 1),
    axis.params = list(vline = TRUE),
    mapping = aes(x = Genome_Size, y = genome, color = Completeness),
    offset = 3.8,
    orientation = "y",
    stat = "identity"
  ) +
  scale_color_gradient(low = "#cf8888", high = "#a2cc87") +
  labs(color = "Genome completeness") +
  theme(legend.text = element_text(size = 8), title = element_text(size = 10), legend.key.size = unit(0.3, "cm"), )

function_tree
```

# Functional ordination
```{r tSNE_function, comment="", echo=FALSE, message=FALSE, warning=FALSE}
# Generate the tSNE ordination
tSNE_function <- Rtsne(X = function_table, dims = 2, check_duplicates = FALSE)
# Plot the ordination
function_ordination <- tSNE_function$Y %>%
  as.data.frame() %>%
  mutate(genome = rownames(function_table)) %>%
  inner_join(genome_metadata, by = "genome") %>%
  rename(tSNE1 = "V1", tSNE2 = "V2") %>%
  select(genome, phylum, tSNE1, tSNE2, Genome_Size) %>%
  ggplot(aes(x = tSNE1, y = tSNE2, color = phylum, size = Genome_Size)) +
  geom_point(shape = 16, alpha = 0.7) +
  scale_color_manual(values = phylum_colors) +
  theme_minimal() +
  labs(color = "Phylum", size = "Genome size") +
  guides(color = guide_legend(override.aes = list(size = 5))) # enlarge Phylum dots in legend

function_ordination
```
