==> src/Markdown/05_RQL_lizards.Rmd <==
---
title: "RLQ lizards"
output: html_document
date: "2024-01-21"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(ade4)
library(tidyverse)
library(vegan)
library(ggplot2)
library(ggrepel)
library(gridExtra)
library(RColorBrewer)
library(distillR)
```

```{r }
load("/Users/jtk712/Documents/Projects/lizard_sample_types/resources/data/data.Rdata")
```

```{r}
# Aggregate bundle-level GIFTs into the compound level
GIFTs_elements <- to.elements(genome_gifts, GIFT_db)
GIFTs_elements_filtered <- GIFTs_elements[rownames(GIFTs_elements) %in% genome_counts$genome, ]
GIFTs_elements_filtered <- as.data.frame(GIFTs_elements_filtered) %>%
  select_if(~ !is.numeric(.) || sum(.) != 0)

# Aggregate element-level GIFTs into the function level
GIFTs_functions <- to.functions(GIFTs_elements_filtered, GIFT_db)
functions <- GIFTs_functions %>%
  as.data.frame()

# Aggregate function-level GIFTs into overall Biosynthesis, Degradation and Structural GIFTs
GIFTs_domains <- to.domains(GIFTs_functions, GIFT_db)
domains <- GIFTs_domains %>%
  as.data.frame()
# Get community-weighed average GIFTs per sample
genome_counts_row <- genome_counts %>% column_to_rownames(., "genome")
GIFTs_elements_community <- to.community(GIFTs_elements_filtered, genome_counts_row, GIFT_db)
GIFTs_functions_community <- to.community(GIFTs_functions, genome_counts_row, GIFT_db)
GIFTs_domains_community <- to.community(GIFTs_domains, genome_counts_row, GIFT_db)
```

```{r elements, comment="", echo=FALSE, message=FALSE, warning=FALSE}
mag_weighted <- round(
  sweep(genome_counts_row, MARGIN = 1, 1000, `*`), 0
)

mag_weighted_t <- mag_weighted %>%
  t() %>%
  #  row_to_names(row_number = 1) %>%
  as.data.frame() %>%
  rownames_to_column(., "samples") %>%
  dplyr::arrange(samples) %>%
  column_to_rownames(., "samples")
```


```{r}
# metadata binary
design <- sample_metadata[, c(
  "sample", "Individual", "Sample_type"
)]
design$Type <- 0
design$Type[design$Sample_type == "feces"] <- 1
for (i in 1:ncol(design)) {
  if (is.character(design[, i])) {
    design[, i] <- factor(design[, i])
  }
}
```

# RLQ analysis (functions)

The multivariate RLQ analysis relates a species x traits matrix (Q, species are rows and columns are biological traits) to a matrix of 
environmental variables at each site (R, sites are rows and columns are environmental descriptors), using a species x abundance matrix 
(L, sites are rows and columsn are abundances of specific species) as a link.

```{r}
genome_counts_t_rel <- decostand(mag_weighted_t, method = "total")
genome_counts_t_log <- log1p(genome_counts_t_rel)
```


```{r}
four <- fourthcorner(data.frame(design[c(4)]), genome_counts_t_log, functions,
  modeltype = 6,
  nrepet = 999, p.adjust.method.G = "fdr", p.adjust.method.D = "fdr"
)
summary(four)
domains
four_domains <- fourthcorner(data.frame(design[c(4)]), genome_counts_t_log, domains,
  modeltype = 6,
  nrepet = 999, p.adjust.method.G = "fdr", p.adjust.method.D = "fdr"
)
summary(four_domains)
```
